refactor: migrate hooks from lib to hooks directory
